move ui-related stuff to ui.py
