+++ src/add_data_advanced.py
@@ -9,7 +9,7 @@ def open_add_data_advanced_window(widget, _, self):
     """
     win = AddAdvancedWindow(self)
     button = win.open_advanced_confirm_button
-    win.set_transient_for(self.props.active_window)
+    win.set_transient_for(self.main_window)
     win.set_modal(True)
     button.connect("clicked", on_accept, self, win)
     win.present()
