+++ src/graphs.py
@@ -3,13 +3,13 @@ import os
 import numpy
 
 from gi.repository import Gtk, Adw
-from . import plotting_tools, samplerow, colorpicker, utilities, file_io
+from . import plotting_tools, samplerow, colorpicker, utilities, file_io, ui
 from .canvas import Canvas
 from .data import Data
 from .misc import DummyToolbar, ImportSettings
 
 def get_theme_color(self):
-    win = self.props.active_window
+    win = self.main_window
     styles = win.get_style_context()
     rgba = styles.lookup_color('theme_bg_color')[1]
     rgba_tuple = (round(rgba.red*255), round(rgba.green*255), round(rgba.blue*255), round(rgba.alpha*255))
@@ -63,7 +63,7 @@ def open_files(self, files):
                             if handle_duplicates == "Auto-rename duplicates":
                                 item.filename = get_duplicate_filename(self, item.filename)
                             elif handle_duplicates == "Ignore duplicates":
-                                self.props.active_window.toast_overlay.add_toast(Adw.Toast(title=f"Item \"{item.filename}\" already exists"))
+                                self.main_window.toast_overlay.add_toast(Adw.Toast(title=f"Item \"{item.filename}\" already exists"))
                                 return
                             elif handle_duplicates == "Override existing items":
                                 y_axis = item.plot_Y_position
@@ -79,8 +79,7 @@ def open_files(self, files):
                 add_sample_to_menu(self, item.filename, item.color, item.key, select_item = True)
     self.canvas.draw()
     plotting_tools.set_canvas_limits_axis(self, self.canvas)
-    plotting_tools.refresh_plot(self)
-    enable_data_dependent_buttons(self, utilities.get_selected_keys(self))
+    toggle_data(None, self)
 
 def get_duplicate_filename(self, name):
     loop = True
@@ -94,19 +93,10 @@ def get_duplicate_filename(self, name):
                 loop = True
     return new_name
 
-def toggle_darkmode(shortcut, theme, widget, self):
-    if Adw.StyleManager.get_default().get_dark():
-        self.plot_settings.plot_style = self.preferences.config["plot_style_dark"]
-    else:
-        self.plot_settings.plot_style = self.preferences.config["plot_style_light"]
-    plotting_tools.reload_plot(self)
-
-
 def select_item(self, key):
     item = self.item_rows[key]
     item.check_button.set_active(True)
-    plotting_tools.refresh_plot(self)
-    enable_data_dependent_buttons(self, utilities.get_selected_keys(self))
+    toggle_data(None, self)
 
 def swap(str1):
     str1 = str1.replace(',', 'third')
@@ -129,7 +119,7 @@ def delete(widget,  self, id, give_toast = True):
     del self.item_rows[id]
     del self.datadict[id]
     if give_toast:
-        self.props.active_window.toast_overlay.add_toast(Adw.Toast(title=f"Deleted {filename}"))
+        self.main_window.toast_overlay.add_toast(Adw.Toast(title=f"Deleted {filename}"))
 
     if len(self.datadict) == 0:
         self.canvas.ax.legend().remove()
@@ -138,22 +128,18 @@ def delete(widget,  self, id, give_toast = True):
         self.main_window.no_data_label_box.set_visible(True)
 
     reset_clipboard(self)
-    plotting_tools.refresh_plot(self)
-    enable_data_dependent_buttons(self, utilities.get_selected_keys(self))
+    toggle_data(None, self)
 
 
 def select_all(widget, _, self):
     for key, item in self.item_rows.items():
         item.check_button.set_active(True)
-    plotting_tools.refresh_plot(self)
-    enable_data_dependent_buttons(self, utilities.get_selected_keys(self))
-
+    toggle_data(None, self)
 
 def select_none(widget, _, self):
     for key, item in self.item_rows.items():
         item.check_button.set_active(False)
-    plotting_tools.refresh_plot(self)
-    enable_data_dependent_buttons(self, False)
+    toggle_data(None, self)
 
 def add_sample_to_menu(self, filename, color, key, select_item = False):
     win = self.main_window
@@ -179,79 +165,11 @@ def add_sample_to_menu(self, filename, color, key, select_item = False):
     row.sample_ID_label.set_text(label)
     self.list_box.append(row)
     self.sample_menu[key] = self.list_box.get_last_child()
+    toggle_data(None, self)
     
-def toggle_data(widget,  self):
+def toggle_data(_, self):
     plotting_tools.refresh_plot(self)
-    enable_data_dependent_buttons(self, utilities.get_selected_keys(self))
-
-def save_file_dialog(self, documenttype="Text file (*.txt)"):
-    def save_file_chooser(action):
-        dialog = Gtk.FileChooserNative.new(
-            title="Save files",
-            parent=self.props.active_window,
-            action=action,
-            accept_label="_Save",
-        )
-        return dialog
-
-    if len(self.datadict) == 1:
-        chooser = save_file_chooser(Gtk.FileChooserAction.SAVE)
-    elif len(self.datadict) > 1:
-        chooser = save_file_chooser(Gtk.FileChooserAction.SELECT_FOLDER)
-
-    if len(self.datadict) == 1:
-        filename = list(self.datadict.values())[0].filename
-        chooser.set_current_name(f"{filename}.txt")
-    try:
-        chooser.set_modal(True)
-        chooser.connect("response", on_save_response, self, False)
-        chooser.show()
-    except UnboundLocalError:
-        self.props.active_window.toast_overlay.add_toast(Adw.Toast(title=f"Could not open save dialog, make sure you have data opened"))
-
-
-def on_save_response(dialog, response, self, project):
-    if response == Gtk.ResponseType.ACCEPT:
-        path = dialog.get_file().peek_path()
-        if project:
-            file_io.save_project(self, path)
-        else:
-            file_io.save_file(self, path)
-
-def save_project_dialog(widget, _, self, documenttype="Graphs Project (*)"):
-    def save_project_chooser(action):
-        dialog = Gtk.FileChooserNative.new(
-            title="Save files",
-            parent=self.props.active_window,
-            action=action,
-            accept_label="_Save",
-        )
-        return dialog
-
-    chooser = save_project_chooser(Gtk.FileChooserAction.SAVE)
-    chooser.set_modal(True)
-    chooser.connect("response", on_save_response, self, True)
-    chooser.show()
-
-
-def open_file_dialog(widget, _, self, open_project):
-    open_file_chooser = Gtk.FileChooserNative.new(
-        title="Open new files",
-        parent=self.props.active_window,
-        action=Gtk.FileChooserAction.OPEN,
-        accept_label="_Open",
-    )
-    open_file_chooser.set_modal(True)
-    open_file_chooser.set_select_multiple(open_project)
-    open_file_chooser.connect("response", on_open_response, self, open_project)
-    open_file_chooser.show()
-
-def on_open_response(dialog, response, self, project):
-    if response == Gtk.ResponseType.ACCEPT:
-        if(project):
-            file_io.load_project(self, dialog.get_files())
-        else:
-            open_files(self, dialog.get_files())
+    ui.enable_data_dependent_buttons(self, utilities.get_selected_keys(self))
 
 def create_data_from_project(self, new_dictionary):
     """
@@ -295,37 +213,10 @@ def load_empty(self):
     self.dummy_toolbar = DummyToolbar(self.canvas)
     win.toast_overlay.set_child(self.canvas)
 
-def disable_clipboard_buttons(self):
-    win = self.main_window
-    win.redo_button.set_sensitive(False)
-    win.undo_button.set_sensitive(False)
-
 def reset_clipboard(self):
     for key, item in self.datadict.items():
         item.xdata_clipboard = [item.xdata]
         item.ydata_clipboard = [item.ydata]
         item.clipboard_pos = -1
-    disable_clipboard_buttons(self)
+    ui.disable_clipboard_buttons(self)
 
-def enable_data_dependent_buttons(self, enabled):
-    win = self.main_window
-
-    dependent_buttons = [
-    win.shift_vertically_button,
-    win.translate_x_button,
-    win.translate_y_button,
-    win.multiply_x_button,
-    win.multiply_y_button,
-    win.smooth_button,
-    win.fourier_button,
-    win.inverse_fourier_button,
-    win.normalize_button,
-    win.center_data_button,
-    win.derivative_button,
-    win.integral_button,
-    win.transform_data_button,
-    win.combine_data_button,
-    ]
-
-    for button in dependent_buttons:
-        button.set_sensitive(enabled)
